<refactor><result实体类移动至各自游戏模式包下>
