+++ app/src/main/java/com/sunzk/colortest/intermediateColor/IntermediateColorPageData.kt
@@ -1,6 +1,5 @@
 package com.sunzk.colortest.intermediateColor
 
-import com.sunzk.colortest.db.bean.IntermediateColorResult
 import com.sunzk.colortest.entity.HSB
 
 /**
@@ -11,10 +10,10 @@ import com.sunzk.colortest.entity.HSB
  * @param randomIndex 左颜色与右颜色的差异点。0=色相，1=饱和度，2=明度,其他值无意义
  */
 data class IntermediateColorPageData(
-    val difficulty: IntermediateColorResult.Difficulty,
-    val questionLeftColor: HSB,
-    val questionRightColor: HSB,
-    val answerColor: HSB,
-    val randomIndex: Int,
-    val hsbBarLock: BooleanArray
+	val difficulty: IntermediateColorResult.Difficulty,
+	val questionLeftColor: HSB,
+	val questionRightColor: HSB,
+	val answerColor: HSB,
+	val randomIndex: Int,
+	val hsbBarLock: BooleanArray
 )
